Remove is_record() from util calls.
This is just done to make the calls more consistent. Previously, in some
places, it was is_record(result_packet), and in others it was
Res = #result_packet. This just makes all calls in util consistent.

Also, I fixed the ordering of the previously introduced functions
insert_id/1 and affected_rows/1, which were originally put in the
-export without being alphabetical. So I made them alphabetical.