Deprecate `emysql_util`, move functionality to `emysql`
To build a nice API, we need to hide a lot of internal modules in the code base. The goal
of this change is to make `emysql` the sole entry point for the module and its use. To retain
backwards compatibility, we still handle calls to `emysql_util`, but this use is on a deprecation
track for the next major release.

While here:

* Fix the `emysql_util_SUITE` so it actually tests something. It would accept any output.
* Begin simplifying the code base by removing stuff that is not needed.